+++ universal-login-commons/lib/types/ContractJSON.ts
