+++ universal-login-commons/lib/types/common.ts
@@ -1,5 +1,3 @@
-import {utils} from 'ethers';
-
 export type Procedure = (...args: any[]) => void;
 
 export type Predicate = (...args: any[]) => boolean;
@@ -7,21 +5,6 @@ export type Predicate = (...args: any[]) => boolean;
 export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
 
 export type PartialRequired<T, TRequired extends keyof T> = Partial<T> & Pick<T, TRequired>;
-
-export interface SignedMessage {
-  gasToken: string;
-  operationType: number;
-  to: string;
-  from: string;
-  nonce: number | string;
-  gasLimit: utils.BigNumberish;
-  gasPrice: utils.BigNumberish;
-  data: utils.Arrayish;
-  value: utils.BigNumberish;
-  signature: string;
-}
-
-export type UnsignedMessage = Omit<SignedMessage, 'signature'>;
 
 export interface Notification {
   id: number;
