+++ universal-login-commons/lib/types/message.ts
@@ -1,0 +1,21 @@
+import {utils} from 'ethers';
+import {Omit, PartialRequired} from './common';
+
+export type Message = Partial<SignedMessage>;
+
+export type MessageWithFrom = PartialRequired<SignedMessage, 'from'>;
+
+export interface SignedMessage {
+  gasToken: string;
+  operationType: number;
+  to: string;
+  from: string;
+  nonce: string | number;
+  gasLimit: utils.BigNumberish;
+  gasPrice: utils.BigNumberish;
+  data: utils.Arrayish;
+  value: utils.BigNumberish;
+  signature: string;
+}
+
+export type UnsignedMessage = Omit<SignedMessage, 'signature'>;
